Use LocalDate/Double, add mapper & transactions
Refactor event DTOs and services: change EventoRequest.fecha from LocalDateTime to LocalDate and Evento.precioEntrada from Integer to Double (added Lombok @Builder). Introduce EventoMapper and switch EventoService to use EventoRequest/EventoResponse instead of old DTO conversion, returning EventoResponse from save and getAll. Mark EventoService and PedidoService as @Transactional and clean up related imports/unused code. These changes separate date/time handling, improve price precision, and centralize mapping logic.

+++ src/main/java/org/example/topdeckapi/src/model/Evento.java
@@ -2,6 +2,7 @@ package org.example.topdeckapi.src.model;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.example.topdeckapi.src.Enumerados.ESTADO_EVENTO;
@@ -34,7 +35,7 @@ public class Evento {
     private LocalTime hora;
 
     @Column(name = "precio_entrada")
-    private Integer precioEntrada;
+    private Double precioEntrada;
 
     @Column(name = "estado")
     @Enumerated(EnumType.STRING)
